chore: seperate db mock data into a seperate file
