+++ server/database/tests/index.test.js
@@ -1,7 +1,7 @@
 import SequelizeMock from 'sequelize-mock';
 import pg from 'pg';
 import { resetAndMockDB } from '@utils/testUtils';
-import { DB_ENV } from '@utils/testUtils/mockData';
+import { DB_ENV } from '@server/utils/testUtils/dbConfig';
 
 const mocks = {};
 describe('getClient', () => {
