+++ server/utils/testUtils/dbConfig.js
@@ -1,0 +1,7 @@
+export const DB_ENV = {
+  DB_URI: 'postgres://reporting_dashboard_role:reportingdashboard123@localhost:5432/reporting_dashboard_dev',
+  POSTGRES_HOST: 'host',
+  POSTGRES_USER: 'user',
+  POSTGRES_PASSWORD: 'password',
+  POSTGRES_DB: 'table'
+};
