+++ server/utils/testUtils/mockData.js
@@ -61,11 +61,3 @@ export const supplierProductsTable = range(1, 10).map((_, index) => ({
   productId: index + 1,
   supplierId: index + 1
 }));
-
-export const DB_ENV = {
-  DB_URI: 'postgres://reporting_dashboard_role:reportingdashboard123@localhost:5432/reporting_dashboard_dev',
-  POSTGRES_HOST: 'host',
-  POSTGRES_USER: 'user',
-  POSTGRES_PASSWORD: 'password',
-  POSTGRES_DB: 'table'
-};
